Parsing rewrite using Prism
Prism is small parser used for code color highlight but it allows
parsing of the tokens as well. It is much more robust than naive bracket
counting used previously. It can be also used for string parsing in
order to manipulate with quotes as well.

This commit is working prototype which needs polishing and then update
for quotes parsing (which has some additional issues)