fix: wrong bracket parsing in string
in case when we wanted to parse brackets but the bracket content was
sting parsing could failed if there was another string on the line. I
have tried to fix this by start parsing from first bracket. It did fix
this issue in most cases. However if we start parsing with selection it
may still evaluate to no brackets found.